% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime.R
\name{spacetime}
\alias{spacetime}
\title{Spatio-temporal maps from incidence objects}
\usage{
spacetime(obj, sf, type = c("map", "heatmap", "none"), breaks = NULL,
  field = "NAME_2", pal = viridis::viridis, ...)
}
\arguments{
\item{obj}{An \code{incidence} objects, as generated by the code{incidence}
package, stratified by geographic units matching that of a provided
shapefile.}

\item{sf}{A shapefile stored as a \code{SpatialPolygonsDataFrame}
object from the \code{sp} package.}

\item{type}{A character string indicating the type of graphics to produce: a
map ("map"), a heatmap ("heatmap"), or no plot ("none").}

\item{breaks}{A numeric vector indicating break points for the color scale.}

\item{field}{A character string indicating the field containing the name of
the spatial units to be used for plotting - the same as the ones used for
incidence stratification.}

\item{pal}{A color palette to be used for plotting.}

\item{...}{Further arguments to be passed to \code{stplot}.}
}
\value{
The function invisibly returns a list containing:
\itemize{
 \item \code{incience_mat}: an incidence matrix
 \item \code{stfdf_obj}: an stfdf object
 \item \code{plot}: the plot produced
}
}
\description{
This function implements similar features to the S4 generic \code{stplot}
from the \code{spacetime} package for \code{incidence} objects (from the
\code{incidence} package), which are stratified by locations.
}
\examples{

\dontrun{
if (require("sp") &&
    require("raster") &&
    require("outbreaks") &&
    require("incidence")) {

## fetch shapefiles for different admin levels
adm2data <- raster::getData('GADM', country = 'SLE', level = 2)
adm1data <- raster::getData('GADM', country = 'SLE', level = 1)
adm0data <- raster::getData('GADM', country = 'SLE', level = 0)

## plot data
plot(adm2data, border = "grey")
plot(adm1data, add = TRUE)
plot(adm0data, add = TRUE, lwd = 2)
text(sp::coordinates(adm2data), labels = adm2data$NAME_2, cex = 0.6)

## make fake data
dat <- ebola_sim$linelist

## randomly assign district location
p <- c(0.15,0.1,0.01,0.01,0.02,0,0,0.05,0.01,0,0.05,0.1,0.25,0.25)
dat$adm2 <- sample(adm2data$NAME_2, size = nrow(dat),
                   replace = TRUE, prob = p)

## make an incidence object
x <- incidence::incidence(dat$date_of_onset, 30, groups = dat$adm2)

## make plot
res <- spacetime(x, adm2data)

## note that the function returns several objects
names(res)
}
}

}
\seealso{
\itemize{
 \item \code{\link[raster]{getData}} to fetch shapefiles freely available from GDAM
 \item the incidence package documented at: \url{http://www.repidemicsconsortium.org/incidence/}
}
}
\author{
Isobel Blake (\email{isobel.blake@imperial.ac.uk})
and Thibaut Jombart (\email{thibautjombart@gmail.com})
}
