% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacecluster.R
\name{spatial_clusters}
\alias{spatial_clusters}
\alias{print.spatial_clusters}
\alias{plot.spatial_clusters}
\title{Identify and plot spatial clusters from point data}
\usage{
spatial_clusters(x, k)

\method{print}{spatial_clusters}(x, ...)

\method{plot}{spatial_clusters}(x, show_clusters = TRUE,
  show_points = FALSE, show_tiles = FALSE, show_density = FALSE,
  pt_col = "red", pt_alpha = 0.3, cl_col = "red", cl_alpha = 0.7,
  dens_alpha = 0.3, dens_bw = 0.005, leg_posi = c(0.1, 0.8), zoom = 13,
  ...)
}
\arguments{
\item{x}{A \code{matrix} oar \code{data.frame} with two columns indicating
longitude and latitude.}

\item{k}{An \code{integer} indicating the number of clusters to be found.q}

\item{show_clusters}{A logical indicating if clusters should be plotted.}

\item{show_points}{A logical indicating if individual points should be
plotted.}

\item{show_tiles}{A logical indicating if Voronoi tesselation should be
plotted.}

\item{show_density}{A logical indicating if case density should be plotted.}

\item{pt_col}{The color to be used for individual points.}

\item{pt_alpha}{The transparency level to be used for individual points.}

\item{cl_col}{The color to be used for clusters.}

\item{cl_alpha}{The transparency level to be used for clusters.}

\item{dens_alpha}{The transparency level to be used for case density.}

\item{dens_bw}{The bandwidth to be used for case density estimation; larger
values correspond to a wider kernel.}

\item{leg_posi}{The (x,y) position of the legend, on a relative scale from 0
to 1.}

\item{zoom}{The zoom level to be used for \code{ggmap}.}

\item{x}{A \code{spatial_clusters} object.}
}
\value{
The function returns a \code{spatial_cluster} object, which is a list with
the following items:

\itemize{
 \item \code{lonlat}: a matrix of lon/lat for individual points
 \item \code{clusters_coords}: a matrix of lon/lat/size for clusters
 \item \code{clusters}: cluster membership for the point locations
 \item \code{voronoi}: voronoi tesselation for the clusters
}
}
\description{
This function identifies spatial clusters from spatial coordinates (longitude
/ latitude) using Kmeans clustering. The resulting object is a
\code{spatial_clusters} object which can be plotted with several options.
}
\examples{

\dontrun{
if (require("outbreaks")) {

lonlat <- ebola_sim$linelist[, c(10, 11)]
head(lonlat)
nrow(lonlat)


res <- spatial_clusters(lonlat, 30)
res
plot(res)

plot(res, show_density = TRUE, show_tiles = TRUE,
     cl_alpha = .2, cl_col = "blue")

}
}
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
